+++ lib/online_migrations/migration.rb
@@ -30,6 +30,15 @@ module OnlineMigrations
     end
     ruby2_keywords(:method_missing) if respond_to?(:ruby2_keywords, true)
 
+    # @private
+    def revert(*args)
+      if command_checker.version_safe?
+        safety_assured { super }
+      else
+        super
+      end
+    end
+
     # Mark a command in the migration as safe, despite using a method that might otherwise be dangerous.
     #
     # @example